requirements.txt
openai==0.27.0
python-dotenv==0.19.2
pyyaml==6.0
pytest==7.1.2
setuptools==41.2.0
wheel==0.43.0